karma tests for flex reorder
